initial

+++ backend/src/main/resources/db/migrations/V1__Initial.sql
@@ -1,6 +1,6 @@
 CREATE TABLE albums
 (
-    id bigint PRIMARY KEY AUTOINCREMENT NOT NULL,
+    id integer PRIMARY KEY AUTOINCREMENT NOT NULL,
     path nvarchar NOT NULL,
     lastCheck bigint
 );
